# open-api/immich-openapi-specs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/asset": {
      "delete": {
        "operationId": "deleteAssets",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetBulkDeleteDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      },
      "post": {
        "operationId": "createAsset",
        "parameters": [
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-immich-checksum",
            "in": "header",
            "description": "sha1 checksum that can be used for duplicate detection before the file is uploaded",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssetMediaDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetMediaResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ],
        "x-immich-lifecycle": {
          "addedAt": "v1.106.0"
        }
      },
      "put": {
        "operationId": "updateAssets",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetBulkUpdateDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
    "/asset/bulk-upload-check": {
      "post": {
        "description": "Checks if assets exist by checksums",
        "operationId": "checkBulkUpload",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetBulkUploadCheckDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetBulkUploadCheckResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
...
    "/asset/exist": {
      "post": {
        "description": "Checks if multiple assets exist on the server and returns all existing - used by background backup",
        "operationId": "checkExistingAssets",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckExistingAssetsDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckExistingAssetsResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
    "/asset/file/{id}": {
      "get": {
        "operationId": "serveFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "isThumb",
            "required": false,
            "in": "query",
            "schema": {
              "title": "Is serve thumbnail (resize) file",
              "type": "boolean"
            }
          },
          {
            "name": "isWeb",
            "required": false,
            "in": "query",
            "schema": {
              "title": "Is request made from web",
              "type": "boolean"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
...
    "/asset/thumbnail/{id}": {
      "get": {
        "operationId": "getAssetThumbnail",
        "parameters": [
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ThumbnailFormat"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
    "/asset/upload": {
      "post": {
        "operationId": "uploadFile",
        "parameters": [
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-immich-checksum",
            "in": "header",
            "description": "sha1 checksum that can be used for duplicate detection before the file is uploaded",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssetDto"
              }
            }
          },
          "description": "Asset Upload Information",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetFileUploadResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ]
      }
    },
...
    "/asset/{id}/file": {
      "get": {
        "operationId": "getOriginalBytes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "isThumb",
            "required": false,
            "in": "query",
            "schema": {
              "title": "Is serve thumbnail (resize) file",
              "type": "boolean"
            }
          },
          {
            "name": "isWeb",
            "required": false,
            "in": "query",
            "schema": {
              "title": "Is request made from web",
              "type": "boolean"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ],
        "x-immich-lifecycle": {
          "addedAt": "v1.106.0"
        }
      },
      "put": {
        "description": "Replace the asset with new file, without changing its id",
        "operationId": "replaceAsset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/AssetMediaReplaceDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetMediaResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ],
        "x-immich-lifecycle": {
          "addedAt": "v1.106.0"
        }
      }
    },
    "/asset/{id}/thumbnail": {
      "get": {
        "operationId": "getThumbnailBytes",
        "parameters": [
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ThumbnailFormat"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "key",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          },
          {
            "api_key": []
          }
        ],
        "tags": [
          "Asset"
        ],
        "x-immich-lifecycle": {
          "addedAt": "v1.106.0"
        }
      }
    },
...
  },
  "info": {
    "title": "Immich",
    "description": "Immich API",
    "version": "1.105.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "Bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "immich_access_token"
      },
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "AssetBulkDeleteDto": {
        "properties": {
          "force": {
            "type": "boolean"
          },
          "ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "ids"
        ],
        "type": "object"
      },
      "AssetBulkUpdateDto": {
        "properties": {
          "dateTimeOriginal": {
            "type": "string"
          },
          "duplicateId": {
            "nullable": true,
            "type": "string"
          },
          "ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "removeParent": {
            "type": "boolean"
          },
          "stackParentId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "ids"
        ],
        "type": "object"
      },
      "AssetBulkUploadCheckDto": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetBulkUploadCheckItem"
            },
            "type": "array"
          }
        },
        "required": [
          "assets"
        ],
        "type": "object"
      },
      "AssetBulkUploadCheckItem": {
        "properties": {
          "checksum": {
            "description": "base64 or hex encoded sha1 hash",
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "checksum",
          "id"
        ],
        "type": "object"
      },
      "AssetBulkUploadCheckResponseDto": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/AssetBulkUploadCheckResult"
            },
            "type": "array"
          }
        },
        "required": [
          "results"
        ],
        "type": "object"
      },
      "AssetBulkUploadCheckResult": {
        "properties": {
          "action": {
            "enum": [
              "accept",
              "reject"
            ],
            "type": "string"
          },
          "assetId": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "reason": {
            "enum": [
              "duplicate",
              "unsupported-format"
            ],
            "type": "string"
          }
        },
        "required": [
          "action",
          "id"
        ],
        "type": "object"
      },
...
      "AssetFileUploadResponseDto": {
        "properties": {
          "duplicate": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "duplicate",
          "id"
        ],
        "type": "object"
      },
...
      "AssetMediaReplaceDto": {
        "properties": {
          "assetData": {
            "format": "binary",
            "type": "string"
          },
          "deviceAssetId": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "fileCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "fileModifiedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "assetData",
          "deviceAssetId",
          "deviceId",
          "fileCreatedAt",
          "fileModifiedAt"
        ],
        "type": "object"
      },
      "AssetMediaResponseDto": {
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AssetMediaStatus"
          }
        },
        "required": [
          "id",
          "status"
        ],
        "type": "object"
      },
      "AssetMediaStatus": {
        "enum": [
          "created",
          "replaced",
          "duplicate"
        ],
        "type": "string"
      },
...
      "CheckExistingAssetsDto": {
        "properties": {
          "deviceAssetIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "deviceId": {
            "type": "string"
          }
        },
        "required": [
          "deviceAssetIds",
          "deviceId"
        ],
        "type": "object"
      },
      "CheckExistingAssetsResponseDto": {
        "properties": {
          "existingIds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "existingIds"
        ],
        "type": "object"
      },
...
      "CreateAssetDto": {
        "properties": {
          "assetData": {
            "format": "binary",
            "type": "string"
          },
          "deviceAssetId": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "fileCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "fileModifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "isOffline": {
            "type": "boolean"
          },
          "isVisible": {
            "type": "boolean"
          },
          "livePhotoData": {
            "format": "binary",
            "type": "string"
          },
          "sidecarData": {
            "format": "binary",
            "type": "string"
          }
        },
        "required": [
          "assetData",
          "deviceAssetId",
          "deviceId",
          "fileCreatedAt",
          "fileModifiedAt"
        ],
        "type": "object"
      },
      "CreateAssetMediaDto": {
        "properties": {
          "assetData": {
            "format": "binary",
            "type": "string"
          },
          "deviceAssetId": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "fileCreatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "fileModifiedAt": {
            "format": "date-time",
            "type": "string"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isFavorite": {
            "type": "boolean"
          },
          "isOffline": {
            "type": "boolean"
          },
          "isVisible": {
            "type": "boolean"
          },
          "sidecarData": {
            "format": "binary",
            "type": "string"
          }
        },
        "required": [
          "assetData",
          "deviceAssetId",
          "deviceId",
          "fileCreatedAt",
          "fileModifiedAt"
        ],
        "type": "object"
      },
...
      "ThumbnailFormat": {
        "enum": [
          "JPEG",
          "WEBP"
        ],
        "type": "string"
      },
...
    }
  }
}